+++ .travis-ci.sh
@@ -55,11 +55,9 @@ fi
 
 export OPAMYES=1
 export OPAMVERBOSE=1
-export CI_CFLAGS=-Werror
+#export CI_CFLAGS=-Werror
 
 eval `opam config env`
-
-opam remote add dev https://github.com/mirage/mirage-dev.git
 
 opam pin add -n mirage-xen-posix .
 opam pin add -n mirage-xen-ocaml .