+++ xen-ocaml/build.sh
@@ -32,39 +32,38 @@ fi
 CC=${CC:-cc}
 PWD=`pwd`
 CFLAGS="-Wall -Wno-attributes ${ARCH_CFLAGS} ${EXTRA_CFLAGS} ${CI_CFLAGS} -DSYS_xen -USYS_linux \
-  -fno-builtin-fprintf -DHAS_UNISTD -Werror=format \
+  -fno-builtin-fprintf -Werror=format \
   $(pkg-config --cflags $PKG_CONFIG_DEPS) \
   "
 
 rm -rf ocaml-src
-cp -r `ocamlfind query ocaml-src` ocaml-src
+cp -r `opam config exec -- ocamlfind query ocaml-src` ocaml-src
 chmod -R u+w ocaml-src
 
 echo Detected OCaml version `ocamlopt -version`
 case `ocamlopt -version` in
-4.01.* | 4.02.[01])
-  echo Applying GC trace patch
-  patch < trace-gc.patch -p 0
-  ;;
-4.03.*)
-  echo Applying OCaml 4.03 config
-  cp config/version-403.h ocaml-src/byterun/caml/version.h
-  patch < clambda-warnings.patch -p 0
-  ;;
-4.04.*)
+4.04.2)
   echo Applying OCaml 4.04 config
   cp config/version-404.h ocaml-src/byterun/caml/version.h
-  patch < clambda-warnings.patch -p 0
-  patch < os-type-xen.patch -p 0
+  BIGARRAY_OBJ="mmap_unix.o"
+  ;;
+4.05.*)
+  echo Applying OCaml 4.05 config
+  cp config/version-405.h ocaml-src/byterun/caml/version.h
+  BIGARRAY_OBJ=""
+  CFLAGS="-D__ANDROID__"
+  ;;
+*)
+  echo unsupported OCaml version `ocamlopt -version`
+  exit 1
+  ;;
 esac
 
+cd ocaml-src && ./configure && cd ..
 cp config/s.h ocaml-src/config/
-cp config/m.${m_file}.h ocaml-src/config/m.h
-cp config/Makefile.${m_file} ocaml-src/config/Makefile
-touch ocaml-src/config/Makefile
 cd ocaml-src
 # cd byterun && make BYTECCCOMPOPTS="${CFLAGS}" BYTECCCOMPOPTS="${CFLAGS}" libcamlrun.a && cd ..
 cd asmrun && make -j${NJOBS} UNIX_OR_WIN32=unix NATIVECCCOMPOPTS="-DNATIVE_CODE ${CFLAGS}" NATIVECCPROFOPTS="-DNATIVE_CODE ${CFLAGS}" libasmrun.a && cd ..
 CFLAGS="$CFLAGS -I../../byterun"
-cd otherlibs/bigarray && make CFLAGS="${CFLAGS}" bigarray_stubs.o mmap_unix.o && cd ../..
-ar rcs libxenotherlibs.a otherlibs/bigarray/bigarray_stubs.o otherlibs/bigarray/mmap_unix.o
+cd otherlibs/bigarray && make CFLAGS="${CFLAGS}" bigarray_stubs.o ${BIGARRAY_OBJ}
+ar rcs ../../libxenotherlibs.a bigarray_stubs.o ${BIGARRAY_OBJ}